# Sprite slice scaling and tiling