… assets/textures/slice_sprite.png
… assets/textures/slice_square.png
… assets/textures/slice_square_2.png
… crates/bevy_sprite/src/texture_slice.rs