# Objective

Implement sprite tiling and [9 slice scaling](https://en.wikipedia.org/wiki/9-slice_scaling) for `bevy_sprite`.
Allowing slice scaling and texture tiling.

Basic scaling vs 9 slice scaling:

![Traditional_scaling_vs_9-slice_scaling](https://user-images.githubusercontent.com/26703856/177335801-27f6fa27-c569-4ce6-b0e6-4f54e8f4e80a.svg)

Slicing example:

<img width="481" alt="Screenshot 2022-07-05 at 15 05 49" src="https://user-images.githubusercontent.com/26703856/177336112-9e961af0-c0af-4197-aec9-430c1170a79d.png">

Tiling example:

<img width="1283" alt="Screenshot 2022-07-06 at 14 35 34" src="https://user-images.githubusercontent.com/26703856/177552197-01ae4b40-9050-4339-9e71-f346749448a1.png">

# Solution

- `Sprite` now has a `draw_mode` field storing a `SpriteDrawMode` enum with three variants:
  - `Simple` 
  - `Tiled` to have sprites tile horizontally and/or vertically
  - `Sliced` allowing 9 slicing the texture and optionally tile some sections with a `Textureslicer`.
- I updated the `bevy_sprite` extraction stage to extract potentially multiple textures instead of one, depending on the `draw_mode`
- I added two examples showcasing the slicing and tiling feature.

# TODO list

- [x] slicing
- [x] scale factor limit
- [x] Slice Tiling:
  - [x]  Basic stretch mode
  - [x] Center tiling
  - [x] Sides tiling
- [x] Sprite tiling without slices

# Blocking
- [ ] ~~Requires https://github.com/bevyengine/bevy/pull/5247 to be merged to move the tile and slice computations in the `prepare` stage of the sprite render pipeline~~
- [ ] https://github.com/bevyengine/bevy/pull/6621
- [ ] Once #5103  is merged I can enable tiling and slicing for texture sheets.
- [ ] Once #5070 is merged I can add this to `bevy_ui`

# To discuss

I want to move the slicer to be an `Asset` and avoid to compute the slices every frame, like for `TextureAtlas` but the slicing depends on the `Sprite::custom_size` and therefore needs to be updated every frame or on `Changed<Sprite>` detection