@ManevilleF 2022-07-05
Now I'm not sure if I can put the slicer in an `Asset`, I can compute the `Rect` values from a given image size and store it but the slices must be computed every frame since the sprite can be *scaled*  or *resized*.

@mockersf how did you do it in https://github.com/vleue/bevy_ninepatch ? I looked at the code after you gave me the link but I don't understand the logic