@ManevilleF 2022-07-06
@alice-i-cecile  I changed the logic to enable sprite tiling even without slicing, using an enum in `Sprite::draw_mode` and removing `TextureSlicer` from being a component.

I think this is better.

But now that I changed a lot of stuff in the extraction stage of the sprite render pipeline, benchmarking is definetely required