@ManevilleF 2022-07-07
The big issue is accessing the `Image` from the handle in the extraction stage. I removed this step for `SpriteDrawMode::Simple` sprites,  meaning no performance regressions.

Now this is still required for `Sliced` and `Tiled` draw modes. A more optimized solution would be to retrieve the image size lazily on the `Added<Sprite>` event or to ask the user to give the image dimensions.
I'm trying to think about a more ergonomic way to handle this, since the image dimensions are accessed again in the `prepare` stage of the pipeline.